requirements.txt
# DeepTeam Red Teaming Framework Requirements

# Core DeepTeam framework
deepteam>=0.2.3

# LLM API clients
anthropic>=0.64.0
groq>=0.31.0
openai>=1.100.0

# Search APIs
tavily-python>=0.3.0

# Google Agent Development Kit (for agentic testing)
google-adk>=1.8.0
google-cloud-aiplatform
google-auth>=2.40.0
litellm

# Web framework (for agent interfaces)
fastapi
uvicorn
httpx

# Data processing and utilities
pandas>=2.3.0
python-dotenv>=1.1.0
pydantic>=2.11.0

# Visualization and reporting
matplotlib>=3.10.0
seaborn>=0.13.0
rich>=14.1.0

# Optional: Streamlit for web UI
streamlit

# Development and testing
pytest
pytest-asyncio